# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 车载视觉感知与人机交互系统 API
  version: 1.0.0
  description: |
    覆盖车牌识别、交警手势识别、车主手势控车、日志监控与告警智能体四个模块。
    Swagger UI：`/api/docs`；原始运行时定义：`/api/openapi.json`。
servers:
  - url: http://{host}:8001
    variables:
      host:
        default: localhost
...
paths:
...
  /api/police-gesture/recognize-video:
    post:
      tags: [交警手势识别]
      summary: 上传长视频识别交警手势
      security: [{bearerAuth: []}]
      parameters:
        - {name: interval, in: query, schema: {type: integer, minimum: 1, maximum: 600, default: 1}}
        - {name: max_results, in: query, schema: {type: integer, minimum: 1, maximum: 300, default: 300}}
        - {name: max_sampled_frames, in: query, schema: {type: integer, minimum: 1, maximum: 5000, default: 900}}
      requestBody: {$ref: '#/components/requestBodies/VideoFile'}
      responses:
        '200': {description: 视频识别完成}
        '500': {$ref: '#/components/responses/RecognitionFailure'}
...
  /api/police-gesture/history:
    get:
      tags: [交警手势识别]
      summary: 查询交警手势识别历史
      parameters: [{$ref: '#/components/parameters/Skip'}, {$ref: '#/components/parameters/HistoryLimit'}]
      responses:
        '200': {description: 历史记录, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GestureHistoryItem'}}}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    Skip: {name: skip, in: query, schema: {type: integer, minimum: 0, default: 0}}
    HistoryLimit: {name: limit, in: query, schema: {type: integer, minimum: 1, default: 20}}
  requestBodies:
...
    VideoFile:
      required: true
      content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary, description: 视频文件}}}}}
  responses:
...
    RecognitionFailure: {description: 图像/模型处理失败, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
...
    GestureHistoryItem: {type: object, properties: {id: {type: integer}, gesture: {type: string}, gesture_cn: {type: string}, confidence: {type: number}, action: {type: string, nullable: true}, annotated_image: {type: string, nullable: true}, created_at: {type: string, format: date-time}}}
...
    Error: {type: object, required: [detail], properties: {detail: {type: string, description: FastAPI 当前实现的错误消息}}}
...